Add option to force parens for mdspan indexing
If set to true it will force the use of kokkos and use parens for
indexing even if std::mdspan and/or __cpp_multidimensional_subscript is
supported by the compiler.
This can make it easier for users of the library if they must support
compilers with and without std::mdspan.